Spectral Indices Computation

Computes one or more spectral indices for an ee$Image or an ee$ImageCollection object.

ee_Image_spectralIndex(
  x,
  index = "NDVI",
  G = 2.5,
  C1 = 6,
  C2 = 7.5,
  L = 1,
  cexp = 1.16,
  nexp = 2,
  alpha = 0.1,
  slope = 1,
  intercept = 0,
  gamma = 1,
  kernel = "RBF",
  sigma = "0.5 * (a + b)",
  p = 2,
  c = 1,
  online = FALSE,
  drop = TRUE
)

ee_ImageCollection_spectralIndex(
  x,
  index = "NDVI",
  G = 2.5,
  C1 = 6,
  C2 = 7.5,
  L = 1,
  cexp = 1.16,
  nexp = 2,
  alpha = 0.1,
  slope = 1,
  intercept = 0,
  gamma = 1,
  kernel = "RBF",
  sigma = "0.5 * (a + b)",
  p = 2,
  c = 1,
  online = FALSE,
  drop = TRUE
)

Arguments

x

An ee$Image or an ee$ImageCollection to compute indices on.

index

Character. Default 'NDVI'. Index or list of indices to compute. The available options are:

vegetation: Compute all vegetation indices.

burn: Compute all burn indices.

water: Compute all water indices.

snow: Compute all snow indices.

drought: Compute all drought indices.

urban: Compute all urban (built-up) indices.

kernel: Compute all kernel indices.

all: Compute all indices listed below.

Check the complete list of indices here.

G

Numeric. Gain factor. It must be set if index is 'EVI'. Default 2.5.

C1

Numeric. Coefficient 1 for the aerosol resistance term. It must be set if index is 'EVI'. Default 6.0.

C2

Numeric. Coefficient 2 for the aerosol resistance term. It must be set if index is 'EVI'. Default 7.5.

L

Numeric. Canopy background adjustment. It must be set if index is 'EVI' or 'SAVI'. Default is 1.0.

cexp

Numeric. It must be set if index is 'OCVI'. Default 1.16.

nexp

Numeric. It must be set if index is 'GDVI'. Default 2.0.

alpha

Numeric. Weighting coefficient. It must be set if index is 'WDRVI'. Default is 0.1.

slope

Numeric. Soil line slope. Default is 1.0.

intercept

Numeric. Soil line slope. Default = 0.0.

gamma

Numeric. Weighting coefficient. It must be set if index is 'ARVI'. Default is 1.0.

kernel

Numeric. It must be set in kernel indices. The available options are:

linear: Linear Kernel.

RBF: Radial Basis Function (RBF) Kernel.

linear: Polynomial Kernel.

The default kernel is 'RBF'.

sigma

Character or Numeric. Length-scale parameter. It must be set in RBF kernel indices. If character, the expression needs to include 'a' and 'b'. If numeric, it must be positive. Default is '0.5 * (a + b)'.

p

Numeric. Kernel degree. It must be set in polynomial kernel indices ('poly'). Default is 2.0.

c

Numeric. Free parameter that trades off the influence of higher-order versus lower-order terms in the polynomial kernel. It must be set in polynomial kernel indices ('poly'). This must be greater than or equal to 0. Default is 1.0.

online

Logical. Whether to retrieve the most recent list of indices directly from the GitHub repository and not from the local copy. Default is FALSE.

drop

Logical. If TRUE, drop the image bands. Default TRUE.

Value

An ee$Image or an ee$ImageCollection with the computed spectral index, or indices, as new bands.

Examples

if (FALSE) {
library(rgee)
library(rgeeExtra)

ee_Initialize()

s2_indices <- ee$ImageCollection("COPERNICUS/S2_SR") %>%
  ee$ImageCollection$first() %>%
  ee$Image$preprocess()%>%
  ee$Image$spectralIndex(c("NDVI", "SAVI"))

names(s2_indices)
# "NDVI" "SAVI"
}